requirements.txt
requests
pandas
numpy
websocket-client
kiteconnect
psycopg2-binary
pymongo
redis
rq
flask
gunicorn
python-dotenv
scipy
arch #For GARCH
TA-Lib  # Technical Analysis.
tensorflow #AI
scikit-learn #AI
python-multipart #For multipart requests.
transformers #Hugging face.
Flask-Cors